refactor: Game 삭제 후 Controller로 기능 이동
